+++ src/browser/provider/built-in/chrome/cdp.js
@@ -1,5 +1,4 @@
 import remoteChrome from 'chrome-remote-interface';
-import browserTools from 'testcafe-browser-tools';
 import { writeFile } from '../../../../utils/promisified-functions';
 
 
@@ -29,10 +28,15 @@ async function setEmulation (runtimeInfo) {
         await client.Network.setUserAgentOverride({ userAgent: config.userAgent });
 
     if (config.touch !== void 0) {
-        await client.Emulation.setTouchEmulationEnabled({
+        const touchConfig = {
             enabled:       config.touch,
             configuration: config.mobile ? 'mobile' : 'desktop'
-        });
+        };
+
+        if (client.Emulation.setEmitTouchEventsForMouse)
+            await client.Emulation.setEmitTouchEventsForMouse(touchConfig);
+        else if (client.Emulation.setTouchEmulationEnabled)
+            await client.Emulation.setTouchEmulationEnabled(touchConfig);
     }
 
     await resizeWindow({ width: config.width, height: config.height }, runtimeInfo);
@@ -83,7 +87,7 @@ export async function takeScreenshot (path, { client, config }) {
 }
 
 export async function resizeWindow (newDimensions, runtimeInfo) {
-    var { browserId, config, viewportSize } = runtimeInfo;
+    var { browserId, config, viewportSize, providerMethods } = runtimeInfo;
 
     var currentWidth  = viewportSize.width;
     var currentHeight = viewportSize.height;
@@ -92,7 +96,7 @@ export async function resizeWindow (newDimensions, runtimeInfo) {
 
 
     if (!config.headless)
-        await browserTools.resize(browserId, currentWidth, currentHeight, newWidth, newHeight);
+        await providerMethods.resizeLocalBrowserWindow(browserId, newWidth, newHeight, currentWidth, currentHeight);
 
     viewportSize.width  = newWidth;
     viewportSize.height = newHeight;