+++ test/server/browser-provider-test.js
@@ -1,7 +1,9 @@
-var expect               = require('chai').expect;
-var Promise              = require('pinkie');
-var testcafeBrowserTools = require('testcafe-browser-tools');
-var browserProviderPool  = require('../../lib/browser/provider/pool');
+var expect                = require('chai').expect;
+var Promise               = require('pinkie');
+var testcafeBrowserTools  = require('testcafe-browser-tools');
+var browserProviderPool   = require('../../lib/browser/provider/pool');
+var parseProviderName     = require('../../lib/browser/provider/parse-provider-name');
+
 
 describe('Browser provider', function () {
     describe('Path and arguments handling', function () {
@@ -131,6 +133,68 @@ describe('Browser provider', function () {
                 .catch(function (error) {
                     expect(error.message).to.contain('Dispose error');
                 });
+        });
+    });
+
+    describe('Browser provider module names handling', function () {
+        it('Should resolve short form of a scoped provider', function () {
+            expect(parseProviderName('@private/package')).to.deep.equal({
+                providerName: '@private/package',
+                moduleName:   '@private/testcafe-browser-provider-package'
+            });
+        });
+
+        it('Should resolve long form of a scoped provider', function () {
+            expect(parseProviderName('@private/testcafe-browser-provider-package')).to.deep.equal({
+                providerName: '@private/package',
+                moduleName:   '@private/testcafe-browser-provider-package'
+            });
+        });
+
+        it('Should resolve short form of a unscoped provider', function () {
+            expect(parseProviderName('package')).to.deep.equal({
+                providerName: 'package',
+                moduleName:   'testcafe-browser-provider-package'
+            });
+        });
+
+        it('Should resolve long form of a unscoped provider', function () {
+            expect(parseProviderName('testcafe-browser-provider-package')).to.deep.equal({
+                providerName: 'package',
+                moduleName:   'testcafe-browser-provider-package'
+            });
+        });
+    });
+
+    describe('Module loading', function () {
+        var dummyProvider = {
+            init: function () {
+                return Promise.resolve();
+            },
+
+            dispose: function () {
+                return Promise.resolve();
+            }
+        };
+
+        before(function () {
+            browserProviderPool.addProvider('@scope/testcafe-browser-provider-dummy', dummyProvider);
+        });
+
+        after(function () {
+            browserProviderPool.removeProvider('@scope/testcafe-browser-provider-dummy');
+        });
+
+        it('Should load scoped browser provider', function () {
+            return browserProviderPool.getProvider('@scope/dummy').then(function (provider) {
+                expect(provider).to.be.not.null;
+            });
+        });
+
+        it('Should load unscoped browser provider', function () {
+            return browserProviderPool.getProvider('chrome').then(function (provider) {
+                expect(provider).to.be.not.null;
+            });
         });
     });
 });