+++ test/server/test-run-commands-test.js
@@ -744,17 +744,17 @@ describe('Test run commands', function () {
 
         it('Should create NavigateToCommand from object', function () {
             var commandObj = {
-                type:     TYPE.navigateTo,
-                url:      'localhost',
-                storages: 'storages'
+                type:          TYPE.navigateTo,
+                url:           'localhost',
+                stateSnapshot: 'stateSnapshot'
             };
 
             var command = createCommand(commandObj);
 
             expect(JSON.parse(JSON.stringify(command))).eql({
-                type:     TYPE.navigateTo,
-                url:      'localhost',
-                storages: 'storages'
+                type:          TYPE.navigateTo,
+                url:           'localhost',
+                stateSnapshot: 'stateSnapshot'
             });
         });
 
@@ -1066,7 +1066,7 @@ describe('Test run commands', function () {
                 actual:        1,
                 message:       null,
 
-                options: { timeout: null }
+                options: {}
             });
         });
 
@@ -1123,9 +1123,7 @@ describe('Test run commands', function () {
                 expected:      1,
                 message:       null,
 
-                options: {
-                    timeout: null
-                }
+                options: {}
             });
         });
     });